+++ docker/clean_active-failover.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+for c in server1 \
+  server2 \
+  server3; do
+  docker rm -f $c
+done